Build ykllvm in `target/<debug|release>/ykllvm`.
Previously we tried building ykllvm in `OUT_DIR` but, unfortunately,
that means when testing we can't access the Cargo variable we defined in
https://github.com/rust-lang/cargo/issues/10927. Previously, things
accidentally worked because we also required PATH to be set to
`/path/to/ykllvm/bin`, but that means you could pick up unexpected
versions of clang/clang++, for example.

This commit rethinks things. It builds a copy of ykllvm in
`target/<debug|release>/ykllvm`. [Technically this breaks Cargo's rules,
but since we're having to worm around a Cargo restriction, I think we
can deal with this: if Cargo's rules change in the future, we can change
the location easily enough.] `yk-config` gains `--cc` and `--cxx` flags
which give the user the path to the C/C++ compiler.

One minor annoyance is that I've had to delete yksmp's tests, as that,
as written, they don't fit into cargo's model. We should bring these
back, but fixing them here feels like one thing too many.

I've tried to keep this commit as minimal as possible: it would
certainly be possible to do various bits of cleaning-up at the same
time, but I think that would obscure what we're most interested in.