We only need one compiler wrapper.
Since we set `YK_COMPILER_PATH` when calling the wrapper, there's no
need to have different wrappers for different compilers. This allows us
to remove a bit of complexity in ccgen too.

While I'm here, I have, self-indulgently, changed the name of the
wrapper. I think `.sh` as a suffix is unhelpful, because a) I don't care
what language an executable file is written in b) I might change that
language later. I also initially called this "wrap-compiler" but that
just felt a bit wrong: it should either be "wrap-compile" or, as I
eventually settled upon as preferable, "compiler-wrapper". The latter is
a bit longer, but feels like it's more likely to be understood.