Remove the libipt backend.
The ykpt decoder is working well for us, and maintaining two backends is
stifling productivity. It has served us well.

This comes with the caveat that we lose the hwtracer_ykpt test suite.
This was a test suite that tested the ykpt backend by using the libipt
backend as an oracle.

Then again, the libipt backend has been disabled for some time (it
doesn't yet track the stack height when traversing unmapped code, so as
to determine if re-encountering mapped code is a return or a callback).

Hopefully our other tests will serve to (indirectly) test the ykpt
decoder, but later we may need to think about how to better test the
ykpt decoder...