The first incarnation of promotion.
To hint to the JIT that a value should be promoted, do:

  x = yk_promote(x);

During tracing, this function observes and records the value of `x`. At
trace compilation time, `x` is replaced by the observed constant.

As discussed with the team, some of the existing inefficiencies of our
JITted code often prevent promotion from being effective. See the
`promote_not_great.c` test for examples.