The first incarnation of variable promotion.
When a function is annotated with `__attribute__((yk_promote(...)))`,
then the named arguments in `...` will be promoted to constants at trace
-compilation time.

E.g.:

```
__attribute__((yk_promote(x,y))
void f(size_t x, size_t y, size_t z) {
  ...
}
```

A call to `f(1, 2, 3)` would have `x` and `y` baked in to the trace as
constants `1` and `2` (and value guards would be inserted).

This change is just the runtime support. There is compiler support in
ykllvm too.