Flatten some of the configuration layers in hwtracer.
Before this commit, there are sort-of 4 layers of configuration
necessary to get a tracer: this more-or-less gets things down to 1 layer
(with the configuration aspect slightly separate).

In some ways this commit is a bit brutal, and more work is needed, but
this seems like a good first step towards simplifying things and making
them more idiomatic Rust.