Directly expose the "tracer" as a trait.
In essence, the basic idea here is to remove `TraceCollectorImpl` which
hid the `dyn Tracer` from a caller, giving them the impression that
there is a single kind of tracer. It's not very clear to me whether this
API could easily have offered different kind of tracers or not: either
way, it's probably not we would expect. This commit is in many ways
crude, because even a seemingly minor internal change percolated
outwards through the API: so this is best viewed as "the minimum thing
which compiles and passes tests" rather than "an absolutely perfect
change in and of itself".