Start to flatten the yktrace type hierarchy.
At first I thought that yktrace's and hwtracer's type hierarchies were
the same (`TracerImpl` and the like), but they're subtly different. I
haven't yet worked out exactly the best way to tidy things up. Because
of that, this commit, like its predecessor, is rather brutal: it
compiles and passes tests; and hopefully is a step on the road to
something nicer. One thing this commit does do is lose hwtracer's
guarantee that it will detect the user trying to run collection twice on
one thread: yktrace previously didn't make that guarantee, so I'm not
sure it makes sense for hwtracer to try and guarantee it.