Push trace decoding selection to the tracer.
Since we may have different types of tracer (various types of hardware
tracers; software tracers; and so on), `MT` can't sensibly offer the
ability to pick different trace decoders -- different backends will have
very different ideas of what decoders they wish to offer.

This commit thus removes the ability to specify a trace decoder within
`MT`, but gives trace decoders a reference to a `Tracer`, so that users
can configure a `Tracer` object and have that later control trace
decoding (if they want to). At the moment, we don't make use of this
facility, because we only have one decoder.