Simplify `Tracer` & `ThreadTracer` in hwtracer.
This more-or-less makes these two traits similar to their namesakes in
`yktrace`. Previously a `ThreadTracer` was passed to a `Tracer`, but
that forces complexity onto the user of the API. By judicious use of
`Arc` (which can be seen as very similar to `Box` in this setting), we
simplify the API. Notably this lets us get rid of the horrible use of
`Any`.

[Since I accidentally created two similar but distinct APIs last week,
I've now had the opportunity to compare the two over time. `yktrace`'s
API is clearly less worse than `hwtracer`'s, so it makes sense to make
the latter look more like the former.]