Change `HotLocationKind::Compiled` from `const *` to `Arc`.
I think the reason we stored this as a raw pointer is because, at one
point, we literally pointed directly at executable machine code: now we
point at a `CompiledTrace` struct which then (indirectly) points at
executable machine code.

In general I'm leery of storing raw pointers when we don't need to do so
because it's easy to get things with raw pointers wrong: in that sense,
simply moving from `const *` to `Arc` is a correctness win. I also
suspect that we'll find it easier to do things like trace expiry by
having access to an `Arc` rather than a raw pointer.

Note that this commit does impose an atomic increment upon each trace
execution: I think this is an inevitable cost (though perhaps in the
future we might move that cost to a sub-struct), so we might as well
bite the bullet now. In terms of our current performance, this cost
is outweighed -- probably by several orders of magnitude! -- by slowness
elsewhere.