Rename the `hwtracer` crate to `perftracer`.
In the past (and there are still bits of those left), the design of
hwtracer-as-was had abstractions intended to allow it to be generic over
different hardware-tracing backends. However, that hasn't quite worked
out as we expected: because the interface to a tracer (PT etc.) goes via
an OS, hwtracer has implicitly ended up being heavily tied to
Linux-perf.

From yk's perspective, this makes little difference: ykrt has its own
abstractions over tracing backends, hopefully allowing it to handle any
hardware or software backend we come across.

So rather than try and make hwtracer internally generic, this commit
bites the bullet and acknowledges that hwtracer is always likely to be a
Linux-perf-only tracer -- and renames it perftracer. This commit
minimally renames the crate and those things that depend on it.