Deoptimise without returning via control point.
After deoptimisation we used to return a pointer to the new stack frames
back to the control point which, after a comparison, ran
`__ykrt_reconstruct_frames` to copy over the new stack. This design
would make it slightly annoying dealing with side-traces as we'd have to
bubble up this pointer all the way back to the control point.

This change allows us to call `__ykrt_reconstruct_frames` immediately
during deoptimisation, jumping to the correct location in the main
interpreter without having to go via the control point. However, in
order to not leak the Arc<CompiledTrace> (which would mean it can never
be garbage collected) we need to pass it into the trace, so that it can
be dropped properly. This has the lucky side-effect that we now have
access to the `CompiledTrace` during a guard failure which makes
implementing side-traces easier.