Adjust to removal of control point loads.
We no longer need to insert loads after the control point at runtime. As it
turns out this simplifies the design of the control point enough to get rid of
a lot of jitmodbuilder code that deals with finding these loads and matching
them with the relevant stores. During deoptimisation, we also had to jump
through some hoops to deoptimise both the values of these loads as well as the
original AOT value they are replacing. This is no longer neccessary so the
code dealing with that can be removed too.