Store `MT`s in an `Arc`.
This makes the lifetimes of `MT` rather easier to reason about because
there's now no chance of dropping an `MT` if there's an active reference
alive. It should also make it easier to pass `MT`s around parts of
the system that might need to store them (and thus break the link with
an `&` reference).