Use a single shared global AOT module.
Yk requires access to the AOT module at various times during execution,
e.g. tracing, deoptimisation. Since accessing an LLVM Module isn't
threadsafe, we used to clone the AOT module for each thread that needed
it, which is slow. We now use a global AOT module that is only loaded
once, and then shared among all threads. The idea is that if two threads
need access to the AOT module at the same time, waiting for the lock to
clear is probably faster than cloning the entire AOT module.

This change also has the advantage that pointers to LLVM values (e.g.
blocks, instructions) stay the same, making it easier in the future to
serialise information into guards for deoptimisation.

Unfortunately, the syntax to to access a `ThreadSafeModule` is a bit
awkward, as it uses `LLVMOrcThreadSafeModuleWithModuleDo` which takes a
function as input that is then called with the locked module. In the
future we may want to consider wrapping a normal `Module` inside a Rust
Mutex, which should allow us simplify this code back to something
similar to what we had before.