Explicitly document where we don't (but should) unwind the stack.
At the moment, the function call in question puts us into UB. Because of
the nature of the containing function, it's probably (but only
probably!) not going to cause us problems right now. But if other things
change later, that might no longer be true, so best to have a reminder
of the danger.