Basic support for outputting JIT stats.
Setting the environment variable `YKD_JITSTATS` to a file path or `-`
(which outputs to stderr) outputs JSON with stats about the yk
compilation process. For example:

```
$ YKD_SERIALISE_COMPILATION=1 YKD_JITSTATS=- src/lua tests/db.lua
testing debug library and debug information
+
testing inspection of parameters/returned values
+
+
testing traceback sizes
testing debug functions on chunk without debug info
OK
{
    "traces_recorded_ok": 11,
    "traces_recorded_err": 0,
    "traces_compiled_ok": 10,
    "traces_compiled_err": 1
}
```

This commit is very much a "first stab to see if this looks like the
sort of thing we want" rather than "here is a complete thing". That
doesn't mean it's not useful already, but it clearly only captures a
tiny amount of the information we might like to see captured.