Record basic timing information in `JITStats`.
This records the time spent in various VM states (compiling, deopting,
interpreting, JIT executing) and shows it in `JITStats` output (in
seconds). These times are "best efforts" rather than "guaranteed to be
perfect". They do, however, already tell a very interesting story:

```
$ YKD_SERIALISE_COMPILATION=1 YKD_JITSTATS=- src/lua tests/db.lua
...
{
    "duration_compiling": 5.5225,
    "duration_deopting": 0.354,
    "duration_jit_executing": 0.2,
    "duration_outside_yk": 0.140,
    "duration_trace_mapping": 3.3757,
    "traces_collected_err": 0,
    "traces_collected_ok": 11,
    "traces_compiled_err": 1,
    "traces_compiled_ok": 10
}
```

The hope is that it's also easy to add new `TimingState` variants: only
two lines of code are needed in `jitstats.rs` to add a new variant.