Cache the stackmap of a compiled trace.
Reading out the entire trace's stackmap each time a guard fails is slow,
especially, since we only need a small part of that stackmap. We could either
modify the stackmap parser to only read the part that is relevant for the
current guard failure, or we read the entire stackmap immediately after
compiling the trace and store it alongside it inside `CompiledTrace`. The latter
has the advantage of moving the stackmap parsing from deopt to inside the
compilation thread.

Storing the result of the parsed stackmap also means we no longer need to keep
the original stackmap data around which can now be freed, squashing another
memory leak that was on our todo list for a while.