Fix FFI calls to LLVMGetThreadSafeModule().
`LLVMGetThreadSafeModule()` is defined in our C++ code and it it called
from both Rust (via the C ABI) and C++ (by the native C++ ABI).

There are two issues with this part of the code.

1) On the C++ side, the bitcode argument is by reference:

```
extern "C" LLVMOrcThreadSafeModuleRef
LLVMGetThreadSafeModule(struct BitcodeSection &Bitcode)
```

I'm not sure it's safe to use C++ references over the C ABI. There's not
much online, but I did find this [0] which suggests that it's a bad
thing to be doing.

2) The definition on the Rust side has a signiture where the bitcode
argument is by-value (as opposed to by reference on the C++ side):

```
extern "C" {
  pub fn LLVMGetThreadSafeModule(bs: BitcodeSection) -> LLVMOrcThreadSafeModuleRef;
}
```

This is certainly incorrect and during development I was getting some
crashes near here, so this could be the cause.

This commit makes `LLVMGetThreadSafeModule()` accept a pointer (not a
reference or a value) in both the C++ and Rust portions of the code.

[0]: https://stackoverflow.com/questions/1906000/c-by-reference-argument-and-c-linkage

All yk tests passing, and yklua tests too.