Add a `Compiler` trait to allow us to have multiple JIT compilers.
This commit minimally adds a `Compiler` trait which will in the future
allow us to swap between different JIT compilers. At the moment this is
slightly farcical because `IRTrace` `compile`s itself, and that
`compile` method assumes an LLVM backend. This commit isn't aiming to
fix that -- though hopefully we'll do so soon! -- but having a trait
seems the right first step towards separating out the different
concerns.