Move more of the LLVM JIT compiler code into `jitc_llvm.rs`.
At a high-level, this commit inlines the call to `irtrace.compile` in:

```rust
impl Compiler for JITCLLVM {
    fn compile(
        &self,
        irtrace: IRTrace,
    ) -> Result<(*const c_void, Option<NamedTempFile>), Box<dyn Error>> {
      irtrace.compile()
    }
}
```

That has various knock-on effects which somewhat drown this out. Mostly
these are rather minor, but note that `llvmbc_section`:

  1. Moves from ykutil to ykrt.
  2. Moves from using linker hacks to `dlsym` (I couldn't get the linker
     hacks to work with the rest of this commit).

Interestingly, this commit indirectly makes clearer the depth of the
assumption that "LLVM is our only JIT", for example in deopting. That
should help us continue to refactor things.