There's no need to store a `Tracer` object.
When we start tracing, all we need to remember is the `ThreadTracer`. If
a `ThreadTracer` (or a mapper) wants access to the `Tracer` (e.g. for
shared caching), it can store that instance itself because the `Tracer`
is `self: Arc<Self>` and can thus store/pass references to itself.

[*Why* this was in the code I no longer remember. I suspect it was
needed at some point in the past and no longer is. But, perhaps, it was
never needed!]