Rename `ykutil` to `ykaddr`.
There are two reasons for this.

The shallow reason is that I find "util" (or "misc") to be names that
enable us to put stuff in them without thinking about where that stuff
should really lie. [At the moment, for example, and despite its name,
ykutil does run-time caching that has correctness implications that we
currently have no easy way of dealing with.]

The deeper reason is that I have gradually moved stuff out of ykutil,
but what remains is (mostly) harder, because it's shared between
hwtracer and ykrt. I'm hoping that as I gradually work on both that it
might become more obvious which way the remaining code in ykutil/ykaddr
should move to (hopefully without duplicating it!). While that
transition is ongoing (who knows how long it'll take...) we don't want
more code to end up in this crate.