Reorder definitions to reflect the order they're used.
Previously structs/funcs/etc in this file were in a rather random order:
now, from top to bottom, they reflect the order that things are used in
a running yk system. This isn't a semantically meaningful change, but it
would definitely have helped me a little bit when I was getting to grips
with this file.