Make a `TraceDecoder` own a trace.
`HWTMapper` previously had this curious signature:

```rust
pub fn map_trace(
    &mut self,
    mut trace_iter: &'a mut dyn Iterator<Item = Result<Block, HWTracerError>>,
) -> Result<Vec<TracedAOTBlock>, HWTracerError> {
```

where the curiosity is the `&'a mut dyn Iterator`. It turns out that's
because a few lines down we have a call to `enumerate`:

```
for (i, block) in &mut trace_iter.enumerate() {
```

If we replace `&mut` with `&` in the function signature, rustc complains
that `enumerate` needs to be used with `&mut`, which is probably why the
function signature is as it is. rustc's complaint is correct, but the
solution it suggests is only one possibility: if you own the thing being
iterated over, you don't need a mutable reference to call `enumerate`.

This commit makes that change. The function signature is now:

```rust
pub fn map_trace(
    &mut self,
    tdec: Box<dyn TraceDecoder>,
) -> Result<Vec<TracedAOTBlock>, HWTracerError> {
```

On its own, this change makes the code mildly more idiomatic, but isn't
a massive change. But, hopefully, it opens the door for future
simplifications.