Allow a trace to decode itself directly.
This moves us from having an external decoder that takes in a trace to
having a trace do its own decoding. This has two advantages:

  1. It's a fairly substantial simplification of code.

  2. It is no longer possible to try using an inappropriate decoder for
     a given trace (for example, I can't use a non-PT decoder on a PT
     trace).

The previous commit slightly obscures the second point. The critical
change in this commit is the removal of this line in `ykrt::hwt`:

```rust
let tdec = default_decoder(self.0).map_err(|_| InvalidTraceError::InternalError)?;
```

What this did was create a decoder for a trace, even though it didn't
know what type of trace it had received.

This commit is not perfect and one thing it highlights is that if we
want to support multiple decoders for a given type of trace, we would
need to have a `TracerBuilder` or similar that would allow the user to
create a tracer (e.g. "PTTracer") and a decoder (e.g. "ykpt") in one
shot. Since we currently have only one tracer and one decoder, we can
punt on that until we need it.