Remove the top-level concept of "decoder".
hwtracer previously assumed that all that tracers would have decoders,
but there are plausible tracers (e.g. a software tracer) that will store
traces in a "full fat" manner that requires no decoding. The idea of
"decoding" as a top-level hwtracer concept therefore doesn't really make
sense, since only some tracers will have decoders.

This commit thus first removes the top-level concept of "decoder".

That, in a sense, orphans `ykpt`, but that's OK, because it has a
natural home in the `pt` module. So we move `ykpt` there, at the same
time flattening its structure a little bit, because it seems unlikely
that `ykpt` needs to be anything more than a decoder.