Remove a `LazyLock`.
Using a `LazyLock` here is overkill, because we can easily perform the
necessary conversion in the `default` function (not that it's worth
measuring this, but it's probably much faster than using the `LazyLock`
which has to do checks for poisoning etc.). While I'm here, use Rust's
`Layout` to do the align calculations rather than do it ourselves.