The `TraceTooLong` error isn't something hwtracer can generate.
This was a hack I put in recently-ish to signal in ykrt that a trace was
too long. My vague memory is that `map_trace` *had* to return a
`HWTracerError` so I put an extra enum in hwtracer to deal with that.

Whether that was necessary or not at the time, it isn't necessary now,
so remove this from hwtracer and keep the error internal to ykrt.