Add a `TracerBuilder` as a front-end for trace construction.
If a particular user has a request for a specific tracer, they can now
make it by selecting a specific `TracerKind` when constructing a
`TracerBuilder`. At the moment I admit that I can't see any use for this
other than selecting different `TracerKind`s, but "builders" are a
common, and fairly easy to extend, Rust idiom, so this seems like a
reasonable step.