Find successful list of prelink and postlink O2 passes.
This commit introduces try_passes.py script, which uses a divide-and-conquer approach to
identify a list of passes that successfully build yklua and execute its tests n times.
Additionally, we added two new arguments to yk-config that allow for the overriding of cflags and
ldflags with custom prelink and postlink flags.