Start work on the new code generator.
We are aiming to replace LLVM as a trace compiler because it's too slow
to get code out the door.

This first unit of work drafts a deserialiser for a (massively
incomplete) AOT IR format. For now we model the top-level IR module,
functions, blocks and instructions.

Currently there are no types, no globals, no constants, only one opcode:
Nop, and there are no instruction arguments yet. Future PRs will expand
the IR as we incrementally add functionality.

The serialisation format is really dumb to start with and we load in
the whole module eagerly. We may need to optimise that later (adding the
ability to only load what we need).

I've added a (temporary) environment variable `YKD_NEW_CODEGEN` to
enable running the tests with the new codegen. Obviously nothing works
yet.

(We can still unit test the IR and new codegen as we go without flipping
the whole test suite over to it)