Reorganise and simplify `HWTracerError`.
Previously `HWTracerError` recorded all sorts of different errors and
forced users of the library to understand them -- which, at least if
they're half as stupid as me, they can't. The challenge for the user is
that some of these errors are transient (e.g. if a trace is interrupted,
simply retrying it is the right thing to do) and some fatal (e.g. you've
supplied incorrect config settings).

This commit rethinks `HWTracerError`, with the main aim being to try and
give the user an easy way of saying "should I retry this or not?" In
particular, the `Temporary` variant is the one that they are most likely
to want to check for specifically.

As part of this, I also realised/decided that some "nice" errors inside
hwtracer are really "todo"s (or "assert"s) in disguise i.e. they suggest
code is incomplete. Percolating those errors outwards seems pointless to
me: if we encounter one, we want to know where it happened, and then we
need to fix hwtracer. So this commit also returns fewer errors full
stop.

To the extent I can manage, this commit is the minimal change necessary.
That means there is one temporary hack in here: `NoMorePackets` is, I
think, really an internal signalling device rather than an actual error.
I need to think whether that's the right API or not, and what might be
an improvement on it.