Implement side-traces.
When a guard fails a specified amount of times (default: 5), we now
compile a side-trace which runs from the guard failure to the control
point. The next time the same guard fails we can now execute the
side-trace instead. There's one caveat: when a side-trace runs all the
way to the end, we currently need to deoptimise back to the control
point, which then runs one iteration in the normal interpreter before
executing a trace again. While this can be fixed by placing one
additional stackmap before the control point, it's probably not
worthwhile as we'll be soon moving to our own trace compiler.