Only convert to a `Weak` `Arc` when we're about to store it.
We use `Weak` to store references to `HotLocation`s in a way that
doesn't cause cycles. The challenge with `Weak` is that the underlying
object could disappear at any point, including before we do an `unwrap`
on it. [In the code as it currently stands, I don't believe that can
happen, but it wouldn't take much refactoring for it to become a
possibility.] This commit thus only downgrades from `Arc` to `Weak` at
the last-possible moment.