Introduce a `GuardId` newtype.
Currently `GuardId`s are `usize`s, which are easily mixed with other
`usize`s, so having a newtype forces us to be explicit about which is
which. This will also make it easier to change the representation of a
`GuardId` later (e.g. to a pointer) if we need to.