Don't use a special value to communicate which is the last guard.
By definition, the last guard in a trace is the one at the end, so
comparing a guard's ID to the number of guards in a trace suffices. This
removes one way we can subtly go wrong, at least in theory (the old
`usize::MAX` issue), and more importantly means we don't have to
communicate a special value between jitmodbuilder and deopt.