Flesh out (very roughly) instruction encoding for Yk IR.
This adds to Yk IR: a few opcodes, an instruction encoding, basic type
encoding, basic constant encoding.

Types and constants are not stored inline inside the instruction
(because types and constants are arbitrarily large, and we don't want
multiple copies of large things hanging around). Instructions instead
store indices to type and constant tables.

Following LLVM's precedent, where a value is identified by the
instruction that generates it, our local variables are a pair: the basic
block index and instruction index of the instruction that creates the
value.

This unit of work has highlighted some problems with using `Display` as
a human-readbale IR formatter:

 - Because constants are not stored inline, to print the actual constant
   we need access to the constant table.

 - Similarly for types.

 - When we print an instruction, if the instruction generates a value,
   we don't have context (from the parent function and block) to print
   the variable's name.

This should be fixable by using a "printer" struct (with state) to do
the printing instead of `Display`.

Another solution would be to have references to the things we need in
`self`, but it's not clear that would be easy.

Sample snippet of human-readable IR at this stage:

```
func main {
  bb0:
    $_ = alloca const[0]
    $_ = call const[1], const[2]
    $_ = call const[3], const[4], $0_0, const[5]
    store $0_1, const[6]
    $_ = alloca const[0]
    $_ = alloca const[0]
    $_ = getelementptr $0_1, const[4]
    $_ = getelementptr $0_1, const[7]
    $_ = getelementptr $0_1, const[8]
    store $0_8, const[6]
    ...
```

As you can see, constants and generated values cannot be printed
properly yet.

I didn't extend our tests just yet, as I want to fix the formatting
first. This will be my next PR.