Check for and fix dead code warnings.
First, we should be checking for dead code warnings. Add defines to CI
instructions for that.

Second, fix the warnings in the tree.

Rust doesn't realise that, although we don't read some of the IR fields,
deku is in fact using them for decoding.

What we can do for now is mark them as `#[deku(temp)]`, which means deku
still consumes them, but excludes them from the struct afterwards. This
means the fields don't exist after decoding, and there are no dead code
warnings.

When we come to actually read these fields, we will be forced to remove
the annotations too, so that's good.