Update dependencies.
Some of these are quite old, and are rather more specific than seems
useful: `x.y.z` only really makes sense when we know for sure that we need
a specific `.z` or `.y.z` version, when generally we don't really know.
So if `.y` (or, when appropriate, `.x`) has increased, I've dropped the
suffix(es), for simplicity's sake.