Add an "IR lowering" test suite.
For each test you feed in textual LLVM IR and it gets compiled with
ykllvm's clang. We then extract and deserialise the Yk IR, and use
lang_tester to match against it.

We do have to be a bit careful when writing tests. Even at -O0, clang
optimises away some LLVM instructions if it deems them redundant. If in
doubt, make instructions depend on dynamic properties (e.g. argc or
argv).

Nonetheless, this is much less painful that manually throwing bytes into
a vector, as we've been doing in unit tests.